Move new notification UI into Vue

+++ CTFd/themes/admin/assets/js/pages/notifications.js
@@ -4,7 +4,10 @@ import $ from "jquery";
 import CTFd from "core/CTFd";
 import { ezQuery, ezAlert } from "core/ezq";
 import Moment from "moment";
-import { htmlEntities } from "core/utils";
+import Vue from "vue/dist/vue.esm.browser";
+import Notification from "../components/notifications/Notification.vue";
+
+const notificationCard = Vue.extend(Notification);
 
 function submit(event) {
   event.preventDefault();
@@ -30,30 +33,16 @@ function submit(event) {
       });
     }
 
-    let date = Moment(response.data.date)
-      .local()
-      .format("MMMM Do, h:mm:ss A");
-    let title = htmlEntities(response.data.title);
-
-    let content = $(`<div class="card bg-light mb-4">
-      <button type="button" data-notif-id="${
-        response.data.id
-      }" class="delete-notification close position-absolute p-3" style="right:0;" data-dismiss="alert" aria-label="Close">
-        <span aria-hidden="true">&times;</span>
-      </button>
-      <div class="card-body">
-        <h3 class="card-title">${title}</h3>
-        <blockquote class="blockquote mb-0">
-          <p>${response.data.content}</p>
-          <small class="text-muted"><span data-time="${
-            response.data.date
-          }">${date}</span></small>
-        </blockquote>
-      </div>
-    </div>`);
-
-    content.find(".delete-notification").click(deleteNotification);
-    $("#notifications-list").prepend(content);
+    let vueContainer = document.createElement("div");
+    $("#notifications-list").prepend(vueContainer);
+    new notificationCard({
+      propsData: {
+        id: response.data.id,
+        title: response.data.title,
+        content: response.data.content,
+        date: response.data.date
+      }
+    }).$mount(vueContainer);
   });
 }
 
@@ -62,17 +51,13 @@ function deleteNotification(event) {
   const $elem = $(this);
   const id = $elem.data("notif-id");
 
-  ezQuery({
-    title: "Delete Notification",
-    body: "Are you sure you want to delete this notification?",
-    success: function() {
-      CTFd.api.delete_notification({ notificationId: id }).then(response => {
-        if (response.success) {
-          $elem.parent().remove();
-        }
-      });
-    }
-  });
+  if (confirm("Are you sure you want to delete this notification?")) {
+    CTFd.api.delete_notification({ notificationId: id }).then(response => {
+      if (response.success) {
+        $elem.parent().remove();
+      }
+    });
+  }
 }
 
 $(() => {
